+++ templates/list-item-not-linked.php
@@ -22,9 +22,8 @@
  */
 
 ?>
-<li class="attachment-<?php echo absint( $file_id ); ?> file_<?php echo esc_attr( $type ); ?> file_<?php echo esc_attr( $subtype ); ?> <?php echo esc_attr( $hide_icon ); ?>">
-								<?php
-									echo esc_html( $attachment->post_title );
-									echo wp_kses_post( $filesize ) . wp_kses_post( $download_button ) . wp_kses_post( $description ) . wp_kses_post( $file_format_label ) . wp_kses_post( $file_date );
-								?>
+<li class="attachment-<?php echo absint( $file_id ); ?> file_<?php echo esc_attr( $type ); ?> file_<?php echo esc_attr( $subtype ); ?> <?php echo esc_attr( $hide_icon ); ?>"><?php
+	echo esc_html( $attachment->post_title );
+	echo wp_kses_post( $filesize ) . wp_kses_post( $download_button ) . wp_kses_post( $description ) . wp_kses_post( $file_format_label ) . wp_kses_post( $file_date );
+?>
 </li>