+++ changelog.md
@@ -10,6 +10,8 @@
 ### Changed
 
 - Optimized iconset handling
+- Optimized style loading of iconsets
+- Updated Block API version to 3
 - Updated dependencies
 
 ## [4.0.1] - 12.11.2025