+++ changelog.md
@@ -11,6 +11,7 @@
 
 - Optimized iconset handling
 - Optimized style loading of iconsets
+- Optimized spaces around the title of downloads without link
 - Updated Block API version to 3
 - Updated dependencies